app/requirements.txt
asgiref==3.8.1
attr==0.3.2
ConfigParser==7.0.0
cryptography==41.0.5
Cython==3.0.10
dl==0.1.0
docutils==0.21.2
Flask_Cors==4.0.0
Flask_JWT_Extended==4.6.0
Flask_Login==0.6.3
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.3
jnius==1.1.0
keyring==25.2.1
lockfile==0.12.2
numpy==2.0.0
Pillow==10.4.0
pyOpenSSL==23.3.0
pyOpenSSL==24.1.0
railroad==0.5.0
Sphinx==7.3.7
supabase==2.5.1
thread==2.0.3
tornado==6.4.1
trove_classifiers==2024.7.1
truststore==0.9.1
urllib3_secure_extra==0.1.0
watchdog==2.3.1
xmlrpclib==1.0.1
